+++ trailscraper/policy_generator.py
@@ -1,5 +1,5 @@
 """Functions responsible for generating a policy from a set of CloudTrail Records"""
-
+import datetime
 import toolz as toolz
 from toolz import pipe
 from toolz.curried import filter as filterz
@@ -25,10 +25,19 @@ def _by_role_arns(arns_to_filter_for):
     return lambda record: (record.assumed_role_arn in arns_to_filter_for) or (len(arns_to_filter_for) == 0)
 
 
-def generate_policy_from_records(records, arns_to_filter_for=None):
+def _by_timeframe(from_date, to_date):
+    return lambda record: record.event_time is None or \
+                          (from_date <= record.event_time <= to_date)
+
+
+def generate_policy_from_records(records,
+                                 arns_to_filter_for=None,
+                                 from_date=datetime.datetime(1970, 1, 1),
+                                 to_date=datetime.datetime.now()):
     """Generates a policy from a set of records"""
 
     statements = pipe(records,
+                      filterz(_by_timeframe(from_date, to_date)),
                       filterz(_by_role_arns(arns_to_filter_for)),
                       mapz(Record.to_statement),
                       _combine_statements_by(lambda statement: statement.Resource),